Fix crash. Make type specific onViewDetachedFromWindow functions
FooterViewHolder must be treated differently than EpisodeViewHolder by
the adapter because when the FooterViewHolder is created it inflates a
view that does not have some of the elements that EpisodeViewHolder
expects to have available, for instance

    mDownloadButton = (DownloadButtonView) view.findViewById(R.id.feedview_download_button);

will be initialized to null in the constructor when creating
FooterViewHolder. This will later result in a crash when the adapter in
onEpisodeViewDetachedFromWindow unconditionally tries to call

    episodeViewHolder.mDownloadButton.enabledProgressListener(false);

since mDownloadButton is null:

--------- beginning of crash
07-11 01:14:32.795 22043-22043/org.bottiger.podcast E/AndroidRuntime: FATAL EXCEPTION: main
Process: org.bottiger.podcast, PID: 22043
java.lang.NullPointerException: Attempt to invoke virtual method 'void org.bottiger.podcast.views.DownloadButtonView.enabledProgressListener(boolean)' on a null object reference
    at org.bottiger.podcast.activities.feedview.FeedViewAdapter.onViewDetachedFromWindow(FeedViewAdapter.java:214)
    ...

Differating between FooterViewHolder and EpisodeViewHolder was originally
done in onBindViewHolder, but were missing everywhere else where view
holder variables were used. I have fixed this by always checking in
this and following commits, extracting the type specific code into its
own function.

The code is perhaps a bit boilerplaty, but this is as robust as it
possibly could be with the type specific functions being completely
orthogonal. The alternative of just adding a test returning if
FooterViewHolder like originally done in onBindViewHolder has a
significantly higher chance of being forgotten (proof: it was already
forgotten everywhere).

Naming convention for view holder variables:

viewHolder - for the generic functions
episodeViewHolder - for EpisodeViewHolder specific functions
footerViewHolder - for FooterViewHolder specific functions